+++ app/Http/Controllers/RoomController.php
@@ -15,21 +15,24 @@ class RoomController extends Controller
     /**
      * Retrieve all Rooms of an Accommodation.
      *
+     * @param  Request  $request
+     * @param  int  $aid
      * @return Collection|Response
      */
     public function index(Request $request, $aid)
     {
         try {
-
-            // Check if the Accommodation exists.
             $accommodation = Accommodation::find($aid);
             if (!$accommodation) {
                 return response()->error(404, 'Accommodation Not Found');
             }
 
-            // TODO: Check manager status
+            $cid = $accommodation->conference()->first()->getKey();
+            $user = $this->isConferenceManager($request, $cid);
+            if (!$user) {
+                return response()->error(403);
+            }
 
-            // Retrieve its Accommodations.
             return $accommodation->rooms()->get();
         } catch (Exception $e) {
             return response()->error();
@@ -39,21 +42,24 @@ class RoomController extends Controller
     /**
      * Create a Room for an Accommodation.
      *
+     * @param  RoomRequest  $request
+     * @param  int  $aid
      * @return Response
      */
     public function store(RoomRequest $request, $aid)
     {
         try {
-
-            // Check if the Accommodation exists.
             $accommodation = Accommodation::find($aid);
             if (!$accommodation) {
                 return response()->error(404, 'Accommodation Not Found');
             }
 
-            // TODO: Check manager status
+            $cid = $accommodation->conference()->first()->getKey();
+            $user = $this->isConferenceManager($request, $cid);
+            if (!$user) {
+                return response()->error(403);
+            }
 
-            // Create the Room.
             $room = new Room($request->all());
             $room->accommodation()->associate($accommodation);
             $room->save();
@@ -67,27 +73,30 @@ class RoomController extends Controller
     /**
      * Update a Room of an Accommodation.
      *
+     * @param  RoomRequest  $request
+     * @param  int  $aid
+     * @param  int  $rid
      * @return Response
      */
     public function update(RoomRequest $request, $aid, $rid)
     {
         try {
-
-            // Check if the Accommodation exists.
             $accommodation = Accommodation::find($aid);
             if (!$accommodation) {
                 return response()->error(404, 'Accommodation Not Found');
             }
 
-            // TODO: Check manager status.
+            $cid = $accommodation->conference()->first()->getKey();
+            $user = $this->isConferenceManager($request, $cid);
+            if (!$user) {
+                return response()->error(403);
+            }
 
-            // Check if the Room exists.
             $room = Room::find($rid);
             if (!$room) {
                 return response()->error(404, 'Room Not Found');
             }
 
-            // Update the Room.
             $room->fill($request->all())->save();
 
             return response()->success();
@@ -99,28 +108,30 @@ class RoomController extends Controller
     /**
      * Delete a Room of an Accommodation.
      *
+     * @param  RoomRequest  $request
+     * @param  int  $aid
+     * @param  int  $rid
      * @return Response
      */
     public function destroy(RoomRequest $request, $aid, $rid)
     {
         try {
-            // Check if the Accommodation exists.
             $accommodation = Accommodation::find($aid);
             if (!$accommodation) {
                 return response()->error(404, 'Accommodation Not Found');
             }
 
-            // TODO: Check manager status.
+            $cid = $accommodation->conference()->first()->getKey();
+            $user = $this->isConferenceManager($request, $cid);
+            if (!$user) {
+                return response()->error(403);
+            }
 
-            // Check if the Room exists.
             $room = Room::find($rid);
             if (!$room) {
                 return response()->error(404, 'Room Not Found');
             }
 
-            // TODO: Check existence of guests.
-
-            // Delete the Room.
             $room->delete();
 
             return response()->success();
